requirements.txt
langchain
scikit-learn
fastapi
uvicorn[standard]
typing
pydantic
python-dotenv
langchain_community
langchain-google-genai
chromadb
langchain_chroma